backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Isabella Guan <isabellaguan2009@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.104.1"
uvicorn = "0.24.0"
passlib = {version = "1.7.4", extras = ["bcrypt"]}
pymongo = "4.6.0"
pydantic = "2.4.2"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
python-multipart = "0.0.6"
python-dotenv = "1.0.0"
httpx = "0.25.0"
pytest = "7.4.3"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
